Basic implementation of offline mode
This basic implementation isn't detecting and prompting
errors, and if it does, it would probalbyu not be much
accurate or usefull.

The function client_process_input should be rewritten
to allow processing of only one command with status.
This would allow to read through a pipe the output
of the protocol and to prompt out the error.

An other way of achieving an accurate diagnostic is to
read the input line by line (add processing of comment?)
to feed the client, read its ouput line by line too.
This method has advantages:

- it correlates the input with the reported error
- client has not to be changed
- comment (or directives) could be added
- clear could be added to recover from errors